recipe-url-api/requirements.txt
openai>=1.75.0
python-dotenv>=1.0.1
requests>=2.32.3
yt-dlp>=2025.3.31
beautifulsoup4>=4.12.3
fastapi>=0.115.12
uvicorn>=0.34.0
pydantic>=2.11.3
